完善了Redis缓存体系；后端审核问题修复（之前保存评论等级默认是0，现在修复成如果本地审核通过就设置为1，如果本地审核不通过并且转接到ai审核就设置成0）；前端页面调整

+++ pojo/src/main/java/com/wang/pojo/dto/VisitorFollowDTO.java
@@ -10,12 +10,12 @@ public class VisitorFollowDTO {
     /**
      * 访客ID
      */
-    private Long visitorId;
+    private Integer visitorId;
 
     /**
      * 作者ID
      */
-    private Long authorId;
+    private Integer authorId;
 
     /**
      * 作者名称
